Do not overwrite existing config.toml and genesis.json

+++ DOCKER/run_node.sh
@@ -1,20 +1,24 @@
-echo "set moniker = $MONIKER"
-echo "set peers = $PEERS"
+if [ ! -f /tendermint/config/config.toml ]; then
+	echo "set moniker = $MONIKER"
+	echo "set peers = $PEERS"
 
-iface=$(route -n | grep '^0.0.0.0' | awk '{print $8}')
-extaddr=$(ip -f inet a show dev $iface | grep '\<inet\>' | head -1 | awk '{print $2}' | awk -F'/' '{print $1}')
-echo "set extaddr = $extaddr"
+	iface=$(route -n | grep '^0.0.0.0' | awk '{print $8}')
+	extaddr=$(ip -f inet a show dev $iface | grep '\<inet\>' | head -1 | awk '{print $2}' | awk -F'/' '{print $1}')
+	echo "set extaddr = $extaddr"
 
-cp config.toml.in config.toml
-sed -e s/@moniker@/$MONIKER/ -i.tmp config.toml
-sed -e s/@peers@/$PEERS/ -i.tmp config.toml
-sed -e s/@external@/tcp:\\/\\/$extaddr:26656/ -i.tmp config.toml
+	cp config.toml.in config.toml
+	sed -e s/@moniker@/$MONIKER/ -i.tmp config.toml
+	sed -e s/@peers@/$PEERS/ -i.tmp config.toml
+	sed -e s/@external@/tcp:\\/\\/$extaddr:26656/ -i.tmp config.toml
 
-mkdir -p /tendermint/config/
-mkdir -p /tendermint/data/
+	mkdir -p /tendermint/config/
+	mkdir -p /tendermint/data/
 
-mv -f config.toml /tendermint/config/
-mv -f genesis.json /tendermint/config/
+	mv -f config.toml /tendermint/config/
+fi
+if [ ! -f /tendermint/config/genesis.json ]; then
+	cp -f genesis.json.sample /tendermint/config/genesis.json
+fi
 
 # val1 == genesis validator
 if [ "$MONIKER" == "val1" ]; then
